refactor(step3): AuthService와 MemberService 의존관계 수정
