+++ README.md
@@ -77,9 +77,17 @@
 # 🚀 3단계 - 패키지 리팩터링
 
 ## 요구사항
-- [ ] 작성한 인수 테스트를 기반으로 리팩터링 진행
+- [x] 작성한 인수 테스트를 기반으로 리팩터링 진행
 
 ## 프로그래밍 요구사항
-- [ ] 인증 관련 부분을 다른 프로젝트에서 재사용 할 수 있도록 리팩터링
+- [x] 인증 관련 부분을 다른 프로젝트에서 재사용 할 수 있도록 리팩터링
   - [x] 일반 로그인 로직과 github 로그인 로직에 대해 추상화 및 중복 코드 제거
   - [x] 패키지 사이의 상호 참조를 없애기
+  ```
+    시도 1. AuthService -> <<MemberService>> <- MemberServiceImpl 구조
+          - MemberService를 인터페이스로 만들어 AuthService가 구현체에 의존하지 않도록 한다.
+    시도 2. AuthService -> <<UserDetailService>> <- UserDetailServiceImpl -> MemberService 구조
+          - MemberService에서 AuthService가 사용하는 메서드를 추출, UserDetailService로 만든다.
+          - UserServiceImpl은 MemberRepository와 MemberService에 의존하여 메서드를 구현한다.     
+  ```
+
