+++ src/main/java/nextstep/auth/application/AuthService.java
@@ -3,8 +3,6 @@ package nextstep.auth.application;
 import nextstep.auth.AuthenticationException;
 import nextstep.auth.application.dto.GithubProfileResponse;
 import nextstep.auth.application.dto.TokenResponse;
-import nextstep.member.application.MemberService;
-import nextstep.member.application.MemberServiceImpl;
 import nextstep.member.application.dto.MemberRequest;
 import nextstep.member.domain.Member;
 import org.springframework.stereotype.Service;
@@ -13,19 +11,19 @@ import java.util.Optional;
 
 @Service
 public class AuthService {
-    private final MemberService memberService;
+    private final UserDetailService userDetailService;
     private final JwtTokenProvider jwtTokenProvider;
     private final GithubClient githubClient;
 
-    public AuthService(MemberServiceImpl memberService, JwtTokenProvider jwtTokenProvider, GithubClient githubClient) {
-        this.memberService = memberService;
+    public AuthService(UserDetailService userDetailService, JwtTokenProvider jwtTokenProvider, GithubClient githubClient) {
+        this.userDetailService = userDetailService;
         this.jwtTokenProvider = jwtTokenProvider;
         this.githubClient = githubClient;
     }
 
     public TokenResponse authenticateWithEmail(String email, String password) {
-        Member member = memberService.findMemberByEmailOrElseThrow(email);
-        if (!member.getPassword().equals(password)) {
+        Member member = userDetailService.findMemberByEmailOrElseThrow(email);
+        if (!member.match(password)) {
             throw new AuthenticationException();
         }
         return createTokenAndResponse(member.getEmail());
@@ -34,10 +32,9 @@ public class AuthService {
     public TokenResponse authenticateWithGithub(String code) {
         String token = githubClient.requestGithubToken(code);
         GithubProfileResponse response = githubClient.requestUserProfile(token);
-        Optional<Member> member = memberService.findMemberByEmail(response.getEmail());
-
+        Optional<Member> member = userDetailService.findMemberByEmail(response.getEmail());
         if (member.isEmpty()) {
-            memberService.createMember(new MemberRequest(response.getEmail(), null, response.getAge()));
+            userDetailService.createMember(new MemberRequest(response.getEmail(), null, response.getAge()));
             return createTokenAndResponse(response.getEmail());
         }
