+++ src/main/java/nextstep/member/application/MemberService.java
@@ -4,22 +4,38 @@ import nextstep.member.application.dto.MemberRequest;
 import nextstep.member.application.dto.MemberResponse;
 import nextstep.member.domain.LoginMember;
 import nextstep.member.domain.Member;
+import nextstep.member.domain.MemberRepository;
+import org.springframework.stereotype.Service;
 
-import java.util.Optional;
+@Service
+public class MemberService {
+    private final MemberRepository memberRepository;
+    public MemberResponse createMember(MemberRequest request) {
+        Member member = memberRepository.save(request.toMember());
+        return MemberResponse.of(member);
+    }
 
-public interface MemberService {
+    public MemberService(MemberRepository memberRepository) {
+        this.memberRepository = memberRepository;
+    }
 
-    MemberResponse createMember(MemberRequest request);
+    public MemberResponse findMember(Long id) {
+        Member member = memberRepository.findById(id).orElseThrow(RuntimeException::new);
+        return MemberResponse.of(member);
+    }
 
-    MemberResponse findMember(Long id);
+    public void updateMember(Long id, MemberRequest param) {
+        Member member = memberRepository.findById(id).orElseThrow(RuntimeException::new);
+        member.update(param.toMember());
+    }
 
-    void updateMember(Long id, MemberRequest param);
+    public void deleteMember(Long id) {
+        memberRepository.deleteById(id);
+    }
 
-    void deleteMember(Long id);
-
-    MemberResponse findMe(LoginMember loginMember);
-
-    Member findMemberByEmailOrElseThrow(String email);
-
-    Optional<Member> findMemberByEmail(String email);
+    public MemberResponse findMe(LoginMember loginMember) {
+        return memberRepository.findByEmail(loginMember.getEmail())
+                .map(it -> MemberResponse.of(it))
+                .orElseThrow(RuntimeException::new);
+    }
 }
