--- src/main/java/nextstep/member/application/MemberServiceImpl.java
@@ -1,52 +1,0 @@
-package nextstep.member.application;
-
-import nextstep.member.application.dto.MemberRequest;
-import nextstep.member.application.dto.MemberResponse;
-import nextstep.member.domain.LoginMember;
-import nextstep.member.domain.Member;
-import nextstep.member.domain.MemberRepository;
-import org.springframework.stereotype.Service;
-
-import java.util.Optional;
-
-@Service
-public class MemberServiceImpl implements MemberService {
-    private final MemberRepository memberRepository;
-
-    public MemberServiceImpl(MemberRepository memberRepository) {
-        this.memberRepository = memberRepository;
-    }
-
-    public MemberResponse createMember(MemberRequest request) {
-        Member member = memberRepository.save(request.toMember());
-        return MemberResponse.of(member);
-    }
-
-    public MemberResponse findMember(Long id) {
-        Member member = memberRepository.findById(id).orElseThrow(RuntimeException::new);
-        return MemberResponse.of(member);
-    }
-
-    public void updateMember(Long id, MemberRequest param) {
-        Member member = memberRepository.findById(id).orElseThrow(RuntimeException::new);
-        member.update(param.toMember());
-    }
-
-    public void deleteMember(Long id) {
-        memberRepository.deleteById(id);
-    }
-
-    public Member findMemberByEmailOrElseThrow(String email) {
-        return memberRepository.findByEmail(email).orElseThrow(RuntimeException::new);
-    }
-
-    public MemberResponse findMe(LoginMember loginMember) {
-        return memberRepository.findByEmail(loginMember.getEmail())
-                .map(it -> MemberResponse.of(it))
-                .orElseThrow(RuntimeException::new);
-    }
-
-    public Optional<Member> findMemberByEmail(String email) {
-        return memberRepository.findByEmail(email);
-    }
-}
