+++ src/main/java/nextstep/member/domain/Member.java
@@ -47,4 +47,8 @@ public class Member {
     public boolean checkPassword(String password) {
         return Objects.equals(this.password, password);
     }
+
+    public boolean match(String password) {
+        return this.password.equals(password);
+    }
 }
