+++ src/main/java/nextstep/member/ui/MemberController.java
@@ -1,7 +1,7 @@
 package nextstep.member.ui;
 
 import nextstep.auth.ui.AuthenticationPrincipal;
-import nextstep.member.application.MemberServiceImpl;
+import nextstep.member.application.MemberService;
 import nextstep.member.application.dto.MemberRequest;
 import nextstep.member.application.dto.MemberResponse;
 import nextstep.member.domain.LoginMember;
@@ -12,11 +12,12 @@ import java.net.URI;
 
 @RestController
 public class MemberController {
-    private MemberServiceImpl memberService;
+    private final MemberService memberService;
 
-    public MemberController(MemberServiceImpl memberService) {
+    public MemberController(MemberService memberService) {
         this.memberService = memberService;
     }
+
 
     @PostMapping("/members")
     public ResponseEntity<Void> createMember(@RequestBody MemberRequest request) {
